Replace std Mutex with tokio Mutex in TunParamsGen
The inner tunnel parameter generator is protected by a mutex. However
the mutex guard is held across an await point in `generate` by the outer
tunnel parameter generator. Holding standard mutexes across await points
is not allowed.

Replaces the std mutex with a tokio mutex which does allow this, however
this requires that all outer tunnel parameter functions are async. This
commit changes those functions and makes sure they are called correctly.

An alternative solution is to put the inner tunnel parameter generator
in its own thread and turn it into an actor. However that would still
require making the public interfaces async.

It is not trivial to put the inner tpg behind a sync interface as it
must have the `generate` function as async since that function
communicates with the account manager actor.
It would be possible to block on that future however it would be a
smelly hack.