Rename ParsedAppVersion::from_str to parse_from_str
Clippy does not like that we call a method `from_str` since there is a
trait called `FromStr` which implements the same method. We made the
decision to not implement `FromStr` since it would change the signature
of the method from returning an Option to a Result and the purpose of
this refactoring is to change as little of the logic as possible.